feat: Sprint 1完成 - 交易所账户配置和市场数据
完成Sprint 1的两个核心需求：
- FR-09: 交易所账户配置管理
- FR-02: 市场数据和自选币种管理

新增功能：
1. 交易所账户管理（加密存储、API验证）
2. 实时市场数据获取（Binance API）
3. 自选币种管理
4. K线数据查询

技术改进：
1. 财务数据类型从Float改为Numeric(20, 8)
2. API Key使用Fernet加密存储
3. 数据库事务处理（try-except-rollback）
4. 添加mypy配置
5. 代码质量检查通过（ruff、mypy）

新增文件：
- 2个文档：AGENTS.md, PROGRESS.md
- 4个数据库迁移
- 19个功能模块文件（模型、API、工具等）

代码统计：+1,903行, -13行

+++ src/backend/app/models/execution.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Enum, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, DateTime, Enum, ForeignKey
 from sqlalchemy.sql import func
 from app.db.database import Base
 import enum
@@ -28,7 +28,7 @@ class StrategyExecution(Base):
     status = Column(Enum(ExecutionStatus), default=ExecutionStatus.RUNNING)
     started_at = Column(DateTime(timezone=True), server_default=func.now())
     ended_at = Column(DateTime(timezone=True))
-    profit = Column(Integer, default=0)
+    profit = Column(Numeric(20, 8), default=0)
     trades_count = Column(Integer, default=0)
     error_message = Column(String(500))
     created_at = Column(DateTime(timezone=True), server_default=func.now())
